requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==3.0.0
Django==5.2.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-health-check==3.19.0
django-redis==5.4.0
django-silk==5.4.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
gprof2dot==2025.4.14
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.25.1
sentry-sdk==2.29.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
